Move Factory method from ViewModel to relevant types & Name it as random(), standard()
